+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 This file will document the changes of PyModMC.
 
+## [Alpha 0.0.2](https://pypi.org/project/PyModMC/0.0.2a0/) - 7/6/2021
+ - Added `FoodItem` function to `Mod` class.
+ - Added timer to `Mod.save` function.
+ - Added docstrings to functions.
+ - Fixed `ModuleNotFoundError` in `__init__.py`.
+
 ## [Alpha 0.0.1](https://pypi.org/project/PyModMC/0.0.1a0/) - 7/1/2021
 Initial release of PyModMC.